Moving ApplyStyle.java to adequate package, updated metadata registry
