+++ org.sysadl.viewpoints/META-INF/MANIFEST.MF
@@ -12,7 +12,8 @@ Require-Bundle: org.eclipse.ui,
  org.sysadl,
  com.google.inject,
  org.eclipse.xtext,
- org.eclipse.sirius.diagram
+ org.eclipse.sirius.diagram,
+ org.sysadl.grammar
 Bundle-ActivationPolicy: lazy
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Import-Package: org.eclipse.ui.ide,
