+++ org.sysadl.viewpoints/plugin.xml
@@ -34,7 +34,7 @@
            id="org.sysadl.viewpoints.addElement">
      </javaActions>
      <javaActions
-           actionClass="org.sysadl.viewpoints.ApplyStyle"
+           actionClass="sysadl.viewpoints.actions.ApplyStyle"
            id="org.sysadl.viewpoints.applyStyle">
      </javaActions>
   </extension>
